Update setup to move original files to backup

+++ setup.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 for f in ${PWD}/.*; do
-  if [[ -f $f ]] ; then
-    echo Copying $(basename "$f" ".")
-    cp "$f" ~/
+  if [ -f $f ] && [ $f != ".DS_Store" ]; then
+    echo Copying $(basename "$f" ".");
+    mv ~/$f ~/$f.bkp;
+    cp "$f" ~/;
   fi
 done
